+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.26.0
+## 0.26.0 (BREAKING CHANGES)
     * use `AsyncCommands` trait from `redis-rs` and remove own `Commands` trait
     * add trait `RedisOps` to support specialized use cases like ping etc.
     * reexport `redis` crate